requirements.txt
qrcode==7.4.2
pillow==9.5.0